Implement authorization using Gates

+++ app/Http/Controllers/QuestionsController.php
@@ -63,6 +63,10 @@ class QuestionsController extends Controller
      */
     public function edit(Question $question)
     {
+        if(\Gate::denies('update-question', $question))
+        {
+            abort(403, 'Access denied ');
+        }  
         return view('questions.edit')->with('question', $question);
     }
 
@@ -75,6 +79,10 @@ class QuestionsController extends Controller
      */
     public function update(AskQuestionRequest $request, Question $question)
     {
+        if(\Gate::denies('update-question', $question))
+        {
+            abort(403, 'Access denied');
+        }
         $question->update($request->only('title', 'body'));
         return redirect()->route('questions.index')->with('success', 'Your question updated successfully.');
     }
@@ -87,6 +95,10 @@ class QuestionsController extends Controller
      */
     public function destroy(Question $question)
     {
+        if(\Gate::denies('delete-question', $question))
+        {
+            abort(403, 'Access denied');
+        }
         $question->delete();
         return redirect()->route('questions.index')->with('success', 'Question has been delete successfully.');
     }
